OneDrive/Documents/tdspro/requirements.txt
fastapi
uvicorn
playwright
google-generativeai
requests
pandas
matplotlib
markdownify
python-dotenv